requirements.txt
openai
qdrant-client
python-dotenv
tiktoken
PyMuPDF
scikit-learn
matplotlib
streamlit
